pyproject.toml
[project]
name = "atomichabit"
version = "0.1.0"
description = ""
authors = [
    {name = "Berry_Hunter",email = "tank-20@list.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (==6.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.2.0,<8.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "eventlet (>=0.40.4,<0.41.0)",
    "drf-yasg (>=1.21.14,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "python-telegram-bot (>=22.6,<23.0)",
    "requests (>=2.32.5,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "django-filter (>=25.2,<26.0)",
    "coverage (>=7.13.4,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''